requirements.txt
mysql-connector-python 
pytest
python-dotenv